+++ trace/table_gtrace.go
@@ -1420,10 +1420,11 @@ func TableOnClose(t *Table, c *context.Context) func(error) {
 		res(p)
 	}
 }
-func TableOnDo(t *Table, c *context.Context, iD string, idempotent bool, nestedCall bool) func(error) func(attempts int, _ error) {
+func TableOnDo(t *Table, c *context.Context, iD string, label string, idempotent bool, nestedCall bool) func(error) func(attempts int, _ error) {
 	var p TableDoStartInfo
 	p.Context = c
 	p.ID = iD
+	p.Label = label
 	p.Idempotent = idempotent
 	p.NestedCall = nestedCall
 	res := t.onDo(p)
@@ -1439,10 +1440,11 @@ func TableOnDo(t *Table, c *context.Context, iD string, idempotent bool, nestedC
 		}
 	}
 }
-func TableOnDoTx(t *Table, c *context.Context, iD string, idempotent bool, nestedCall bool) func(error) func(attempts int, _ error) {
+func TableOnDoTx(t *Table, c *context.Context, iD string, label string, idempotent bool, nestedCall bool) func(error) func(attempts int, _ error) {
 	var p TableDoTxStartInfo
 	p.Context = c
 	p.ID = iD
+	p.Label = label
 	p.Idempotent = idempotent
 	p.NestedCall = nestedCall
 	res := t.onDoTx(p)